[ResourceWatcher] moved tests into component folder
